requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib
bcrypt
python-multipart
email-validator
typing-extensions
nltk
requests
python-dotenv
